# Quick improvements for filterDescendants