--- a/src/models/block.js
+++ b/src/models/block.js
@@ -109,9 +109,7 @@ class Block extends new Record(DEFAULTS) {
    */
 
   get text() {
-    return this.nodes
-      .map(node => node.text)
-      .join('')
+    return this.getText()
   }
 
 }