--- a/src/models/inline.js
+++ b/src/models/inline.js
@@ -109,9 +109,7 @@ class Inline extends new Record(DEFAULTS) {
    */
 
   get text() {
-    return this.nodes
-      .map(node => node.text)
-      .join('')
+    return this.getText()
   }
 
 }