Avoid doing concat on immutable `List` and instead build an array then return it.

It also adds a `forEachDescendant` when we just want to iterate over nodes without creating a return list.