@SamyPesse 2016-11-03
### [With this PR](https://travis-ci.org/GitbookIO/slate/jobs/172898070):

```
Benchmarks
   delete-backward-deep
     Current:	 263 ops/sec (107 runs sampled)
     rme: ±6.36%
   delete-backward-normal
     Current:	 328 ops/sec (109 runs sampled)
     rme: ±6.26%
   delete-backward-wide
     Current:	 340 ops/sec (97 runs sampled)
     rme: ±4.23%
   first-render-normal
     Current:	 64.97 ops/sec (120 runs sampled)
     rme: ±4.22%
   insert-text-normal
     Current:	 926 ops/sec (95 runs sampled)
     rme: ±4.48%
   memoize-util
     Current:	 8,478 ops/sec (86 runs sampled)
     rme: ±8.33%
   normalize-document-deep
     Current:	 1,275 ops/sec (88 runs sampled)
     rme: ±8.98%
   normalize-document-normal
     Current:	 2,671 ops/sec (84 runs sampled)
     rme: ±6.87%
   normalize-document-wide
     Current:	 1,135 ops/sec (83 runs sampled)
     rme: ±4.91%
   render-split-block
     Current:	 27.99 ops/sec (111 runs sampled)
     rme: ±2.81%
   split-block-deep
     Current:	 253 ops/sec (96 runs sampled)
     rme: ±6.42%
   split-block-normal
     Current:	 416 ops/sec (94 runs sampled)
     rme: ±5.76%
   split-block-wide
     Current:	 338 ops/sec (88 runs sampled)
     rme: ±7.75%
```

### [Previously](https://travis-ci.org/ianstormtaylor/slate/jobs/172893921)

```
Benchmarks
   delete-backward-deep
     Current:	 249 ops/sec (107 runs sampled)
     rme: ±4.61%
   delete-backward-normal
     Current:	 318 ops/sec (110 runs sampled)
     rme: ±4.83%
   delete-backward-wide
     Current:	 306 ops/sec (96 runs sampled)
     rme: ±4.73%
   first-render-normal
     Current:	 65.58 ops/sec (123 runs sampled)
     rme: ±3.39%
   insert-text-normal
     Current:	 777 ops/sec (97 runs sampled)
     rme: ±6.27%
   memoize-util
     Current:	 3,409 ops/sec (97 runs sampled)
     rme: ±5.25%
   normalize-document-deep
     Current:	 1,346 ops/sec (90 runs sampled)
     rme: ±5.71%
   normalize-document-normal
     Current:	 3,006 ops/sec (85 runs sampled)
     rme: ±4.99%
   normalize-document-wide
     Current:	 1,320 ops/sec (83 runs sampled)
     rme: ±7.69%
   render-split-block
     Current:	 30.32 ops/sec (104 runs sampled)
     rme: ±6.02%
   split-block-deep
     Current:	 222 ops/sec (101 runs sampled)
     rme: ±4.17%
   split-block-normal
     Current:	 367 ops/sec (95 runs sampled)
     rme: ±7.16%
   split-block-wide
     Current:	 278 ops/sec (92 runs sampled)
     rme: ±5.79%
```