@SamyPesse 2016-11-03
Here is the comparaison between my last 2 commits:

```
Benchmarks

   delete-backward-deep
     Current:	 299 ops/sec (113 runs sampled)
     Reference:	 305 ops/sec (108 runs sampled)
     comparison: Indeterminate
     diff: -1.94%
     rme: ±3.23%

   delete-backward-normal
     Current:	 376 ops/sec (116 runs sampled)
     Reference:	 369 ops/sec (113 runs sampled)
     comparison: Indeterminate
     diff: +2.07%
     rme: ±2.50%

   delete-backward-wide
     Current:	 354 ops/sec (99 runs sampled)
     Reference:	 362 ops/sec (100 runs sampled)
     comparison: Indeterminate
     diff: -2.15%
     rme: ±4.16%

   first-render-normal
     Current:	 83.48 ops/sec (126 runs sampled)
     Reference:	 79.54 ops/sec (124 runs sampled)
     comparison: Faster
     diff: +4.95%
     rme: ±3.23%

   insert-text-normal
     Current:	 1,046 ops/sec (97 runs sampled)
     Reference:	 1,040 ops/sec (96 runs sampled)
     comparison: Faster
     diff: +0.59%
     rme: ±4.54%

   memoize-util
     Current:	 4,849 ops/sec (98 runs sampled)
     Reference:	 4,695 ops/sec (101 runs sampled)
     comparison: Indeterminate
     diff: +3.28%
     rme: ±8.31%

   normalize-document-deep
     Current:	 1,935 ops/sec (89 runs sampled)
     Reference:	 1,996 ops/sec (88 runs sampled)
     comparison: Indeterminate
     diff: -3.06%
     rme: ±5.63%

   normalize-document-normal
     Current:	 3,969 ops/sec (85 runs sampled)
     Reference:	 4,180 ops/sec (86 runs sampled)
     comparison: Indeterminate
     diff: -5.06%
     rme: ±6.97%

   normalize-document-wide
     Current:	 1,825 ops/sec (84 runs sampled)
     Reference:	 2,049 ops/sec (84 runs sampled)
     comparison: Indeterminate
     diff: -10.95%
     rme: ±6.00%

   render-split-block
     Current:	 43.32 ops/sec (116 runs sampled)
     Reference:	 39.68 ops/sec (110 runs sampled)
     comparison: Faster
     diff: +9.19%
     rme: ±5.31%

   split-block-deep
     Current:	 336 ops/sec (101 runs sampled)
     Reference:	 270 ops/sec (95 runs sampled)
     comparison: Faster
     diff: +24.11%
     rme: ±10.00%

   split-block-normal
     Current:	 559 ops/sec (96 runs sampled)
     Reference:	 495 ops/sec (99 runs sampled)
     comparison: Faster
     diff: +12.96%
     rme: ±6.73%

   split-block-wide
     Current:	 442 ops/sec (90 runs sampled)
     Reference:	 391 ops/sec (91 runs sampled)
     comparison: Faster
     diff: +13.14%
     rme: ±8.15%
```

I think the main benefit will be in react update.